Conflicts:
	kifi-backend/modules/common/app/com/keepit/scraper/extractor/Extractor.scala

+++ tools/eddie/eddie-deploy.py
@@ -6,6 +6,47 @@ import boto.ec2
 import spur
 import os
 import time
+from boto.s3.connection import S3Connection
+from boto.s3.key import Key
+import datetime
+
+class S3Asset(object):
+
+  def __init__(self, key):
+    self.key = key
+    self.name = key.name.split(".")[0]
+    (kind, date, time, _, chash) = self.name.split("-")
+    self.serviceType = kind
+    self.hash = chash
+    self.timestamp = datetime.datetime.strptime(date+time, "%Y%m%d%H%M")
+
+
+class S3Assets(object):
+
+  def __init__(self, serviceType, bucket):
+    conn = S3Connection("AKIAINZ2TABEYCFH7SMQ", "s0asxMClN0loLUHDXe9ZdPyDxJTGdOiquN/SyDLi")
+    bucket = conn.get_bucket(bucket)
+    allKeys = bucket.list()
+    self.assets = []
+    for key in allKeys:
+      asset = S3Asset(key)
+      if asset.serviceType==serviceType:
+        self.assets.append(asset)
+
+
+  def latest(self):
+    inOrder = sorted(self.assets, key = lambda x: x.timestamp)
+    if not inOrder:
+      raise Exception("No Version Available!")
+    return inOrder[-1]
+
+  def byHash(self, chash):
+    versions = [a for a in self.assets if a.hash==chash]
+    if not versions:
+      raise Exception("Unknown commit hash: " + chash)
+    return versions[0]
+
+
 
 class ServiceInstance(object):
   def __init__(self, instance):
@@ -38,7 +79,7 @@ def message_hipchat(msg):
   }
   requests.post("https://api.hipchat.com/v1/rooms/message?format=json&auth_token=47ea1c354d1df8e90f64ba4dc25c1b", data=data)
 
-def log(msg): #ZZZ hipchat with notification
+def log(msg):
   amsg = "[" + getpass.getuser() + "] " + msg
   print amsg
   message_irc(amsg)
@@ -85,27 +126,69 @@ if __name__=="__main__":
   else:
     instances = [instance for instance in instances if instance.service==args.serviceType]
 
-  log("Triggered deploy of %s to %s in %s mode" % (args.serviceType.upper(), str([str(inst.name) for inst in instances]), args.mode))
+  assets = S3Assets(args.serviceType, "fortytwo-builds")
 
   command = ["python", "/home/fortytwo/eddie/eddie-self-deploy.py"]
 
+  assets = S3Assets(args.serviceType, "fortytwo-builds")
+
+  version = 'latest'
+  full_version = 'latest'
   if args.version:
-    command.append(args.version)
+    version = args.version
+    try:
+      full_version = assets.byHash(version).name
+    except:
+      full_version = version
+  else:
+    version = assets.latest().hash
+    full_version = assets.latest().name + " (latest)"
+
+  command.append(version)
 
   if args.mode and args.mode=="force":
     command.append("force")
 
-  for instance in instances:
-    shell = spur.SshShell(hostname=instance.ip,username="fortytwo", missing_host_key=spur.ssh.MissingHostKey.warn)
-    remoteProc = shell.spawn(command, store_pid=True, stdout=sys.stdout)
+  log("Triggered deploy of %s to %s in %s mode with version %s" % (args.serviceType.upper(), str([str(inst.name) for inst in instances]), args.mode, full_version))
+
+  if args.mode and args.mode=="force":
     try:
-      while remoteProc.is_running():
-        time.sleep(0.1)
-      remoteProc.wait_for_result()
+      inpt = raw_input("Warning: Force Mode! You sure? [Y,N,Ctrl+C]")
+      if inpt=="Y":
+        remoteProcs = []
+        for instance in instances:
+          shell = spur.SshShell(hostname=instance.ip,username="fortytwo", missing_host_key=spur.ssh.MissingHostKey.warn)
+          remoteProcs.append(shell.spawn(command, store_pid=True, stdout=sys.stdout))
+        log("Deploy Triggered on all instances. Waiting for them to finish.")
+        try:
+          for remoteProc in remoteProcs:
+            while remoteProc.is_running():
+              time.sleep(0.1)
+            remoteProc.wait_for_result()
+        except KeyboardInterrupt:
+          log("Manual Abort. Might be too late (force mode).")
+          for remoteProc in remoteProcs:
+            try:
+              remoteProc.send_signal(2)
+            except:
+              pass
+          sys.exit(1)
+      else:
+        log("Manual Abort.")
     except KeyboardInterrupt:
       log("Manual Abort.")
-      remoteProc.send_signal(2)
-      sys.exit(1)
-
+  else:
+    for instance in instances:
+      shell = spur.SshShell(hostname=instance.ip,username="fortytwo", missing_host_key=spur.ssh.MissingHostKey.warn)
+      remoteProc = shell.spawn(command, store_pid=True, stdout=sys.stdout)
+      try:
+        while remoteProc.is_running():
+          time.sleep(0.1)
+        remoteProc.wait_for_result()
+      except KeyboardInterrupt:
+        log("Manual Abort.")
+        remoteProc.send_signal(2)
+        sys.exit(1)
+    log("Deployment Complete")
